move constants, type conversion and utils into shared space
