+++ local_server/data_anndata/anndata_adaptor.py
@@ -13,7 +13,7 @@ from local_server.common.colors import convert_anndata_category_colors_to_cxg_ca
 from local_server.common.constants import Axis, MAX_LAYOUTS
 from local_server.common.corpora import corpora_get_props_from_anndata
 from local_server.common.errors import PrepareError, DatasetAccessError, FilterError
-from local_server.common.utils.type_conversion_utils import get_schema_type_hint_of_array
+from shared_utils.utils.type_conversion_utils import get_schema_type_hint_of_array
 from local_server.compute.scanpy import scanpy_umap
 from local_server.data_common.data_adaptor import DataAdaptor
 from local_server.data_common.fbs.matrix import encode_matrix_fbs
