+++ local_server/data_common/data_adaptor.py
@@ -9,7 +9,7 @@ from server_timing import Timing as ServerTiming
 from local_server.common.config.app_config import AppConfig
 from local_server.common.constants import Axis
 from local_server.common.errors import FilterError, JSONEncodingValueError, ExceedsLimitError
-from local_server.common.utils.utils import jsonify_numpy
+from shared_utils.utils.utils import jsonify_numpy
 from local_server.data_common.fbs.matrix import encode_matrix_fbs
