+++ server/cli/launch.py
@@ -12,7 +12,7 @@ from server.default_config import default_config
 from server.app.app import Server
 from server.common.config.app_config import AppConfig
 from server.common.errors import DatasetAccessError, ConfigurationError
-from server.common.utils.utils import sort_options
+from shared_utils.utils.utils import sort_options
 
 DEFAULT_CONFIG = AppConfig()
