+++ server/common/annotations/hosted_tiledb.py
@@ -10,7 +10,7 @@ from flask import current_app
 from server.common.annotations.annotations import Annotations
 from server.common.errors import AnnotationCategoryNameError
 from server.common.utils.sanitization_utils import sanitize_values_in_list
-from server.common.utils.type_conversion_utils import get_dtypes_and_schemas_of_dataframe, get_dtype_of_array
+from shared_utils.utils.type_conversion_utils import get_dtypes_and_schemas_of_dataframe, get_dtype_of_array
 from server.db.cellxgene_orm import Annotation
