+++ server/data_common/data_adaptor.py
@@ -8,7 +8,7 @@ from server_timing import Timing as ServerTiming
 from server.common.config.app_config import AppConfig
 from server.common.constants import Axis
 from server.common.errors import FilterError, JSONEncodingValueError, ExceedsLimitError
-from server.common.utils.utils import jsonify_numpy
+from shared_utils.utils.utils import jsonify_numpy
 from server.data_common.fbs.matrix import encode_matrix_fbs
