+++ server/data_cxg/cxg_adaptor.py
@@ -12,8 +12,8 @@ import server.compute.diffexp_cxg as diffexp_cxg
 from server.common.constants import Axis
 from server.common.errors import DatasetAccessError, ConfigurationError
 from server.common.immutable_kvcache import ImmutableKVCache
-from server.common.utils.type_conversion_utils import get_schema_type_hint_from_dtype
-from server.common.utils.utils import path_join
+from shared_utils.utils.type_conversion_utils import get_schema_type_hint_from_dtype
+from shared_utils.utils.utils import path_join
 from server.data_common.data_adaptor import DataAdaptor
 from server.data_common.fbs.matrix import encode_matrix_fbs
 from server.data_cxg.cxg_util import pack_selector_from_mask
