+++ server/test/unit/common/utils/test_type_conversion_utils.py
@@ -5,7 +5,7 @@ from unittest.mock import patch
 import numpy as np
 from pandas import Series, DataFrame
 
-from server.common.utils.type_conversion_utils import (
+from shared_utils.utils.type_conversion_utils import (
     can_cast_to_float32,
     can_cast_to_int32,
     get_dtype_of_array,
