+++ shared_utils/utils/corpora_constants.py
